blue: Fix the way dot call stacks work
This was breaking a bunch and should hopefully work much better now.

Each time a call expression is made we push nil onto the stack and
each time applyFunction is called we pop the top item to see if its not nil.

If its not its the arg that should be used for the call and if it is we
continue as normal.

The tests all pass so this should hopefully work with more complicated
call stacks.

Types were removed from the lexer to make way for map() function but
that has not really been added yet.