blue: add md_to_html, add sanitize_and_minify, add BASIC template support
Also support HTML sending <head> when content wrapped in <html></html> and
its valid.

Template is VERY BASIC but uses blue inside <%= %> or <% %> (inject/dont-inject)
Theres probably some issues with it.
`ctx` is reserved for the map passed in but as long as its not immutable
it will get replaced with its original value after template execution.

Still more to do!