blue: fixup repl
now we can load and save files in the repl.

filebuf will append all the successfully lines run in addition
to any of the loaded files.

this allows the user to then save it all to a file.

eventually it may make sense to replace a file's content if we
load in the same file again.