blue: Finish up help() work
now docstring comments via ## can be used at the top of a module
or the first line of a function's block to append to the help
string that will be returned when calling help(OBJECT).

eventually we want to include public variables as well
and we still need to add doc comments to everything currently
in std and core.