blue: eval: add buffer of 1 to channel
this may not be strictly necessary but I noticed a sort of deadlock
where I was blocked sending (and it should have been receiving?)

So this fixes that issue for now but it may change soon enough once I get
a better understanding.

I also cleaned up the way we call 'self()' so it works in the spawned
function.

NOTE: with the new test added it more than doubled the test run time