blue: allow builtin function name to be used as map key
This annoyed me when I couldnt use the name `wait` for a builtin
(because of go_metrics() returning wait as one of its keys)
So now its supported but you can still shoot yourself in the foot if you
try to call a builtin with a dot call, thats in the map.

ie. var x = {};
x.print = 'something'
x.print() => Not a function NULL
print(x) => still works